fix: Change days_onsite_per_week from i32 to f32 to support fractional days
The LLM extraction was returning fractional values like 2.5 for hybrid work
policies (e.g., "2.5 days per week onsite") but the schema only accepted
integers, causing parsing failures and fallback to regex extraction.

**Change:**
- RemoteWorkDetails.days_onsite_per_week: Option<i32> → Option<f32>

**Test Results:**
✅ 36KB Dice email now extracts via LLM (not regex fallback)
✅ extraction_method: "llm" (was "regex" before)
✅ confidence: 0.85 (high confidence)
✅ Extracted 13 technologies in tech_stack
✅ Full job details: company, industry, location, remote policy all correct

**Impact:**
This completes the full resolution of ISSUE-001. The combination of:
1. Mozilla Readability HTML preprocessing (89.7% size reduction)
2. Float support for fractional days

...enables successful LLM extraction from HTML-heavy recruiter emails.

+++ backend/src/main.rs
@@ -337,7 +337,7 @@ pub struct EmploymentDetails {
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct RemoteWorkDetails {
     pub policy: Option<String>,
-    pub days_onsite_per_week: Option<i32>,
+    pub days_onsite_per_week: Option<f32>,  // Changed to f32 to support fractional days (e.g., 2.5 days/week)
     pub remote_eligible_states: Option<Vec<String>>,
     pub timezone_requirement: Option<String>,
 }
